src/Task_Allocation/requirements.txt
langchain
langchain_community
langchain_core
python-dotenv==1.0.1
langchain_google_genai
pygame
tensorflow
grpcio
customtkinter
